% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{acceptance_check}
\alias{acceptance_check}
\title{Acceptance rate}
\usage{
acceptance_check(
  first.adapt.phase_net,
  round_check,
  last_iter_check,
  prob_mbr,
  layers_def,
  parent_set_combinations,
  BGe_score_all_configs_node,
  omics,
  annot
)
}
\description{
\code{acceptance_check} This phase verify if the acceptance is in range of 0.28 and 0.6.
}
