% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{B_prior_mat}
\alias{B_prior_mat}
\title{Biological prior matrix definition}
\usage{
B_prior_mat(
  omics,
  PK,
  layers_def,
  TFtargs,
  annot,
  lm_METH,
  r_squared_thres,
  p_val_thres,
  TFBS_belief,
  nonGE_belief,
  woPKGE_belief
)
}
\description{
Biological prior matrix definition
}
