% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{BGe_node}
\alias{BGe_node}
\title{BGe score for specific node}
\usage{
BGe_node(
  node,
  adjacency_matrix,
  parent_set_combinations,
  BGe_score_all_configs_node
)
}
\description{
\code{BGe_node} Computes the BGe score of given node using precomputed sets of possible parents.
}
