% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{BGe_score}
\alias{BGe_score}
\title{BGe score}
\usage{
BGe_score(
  adjacency_matrix,
  omics,
  layers_def,
  parent_set_combinations,
  BGe_score_all_configs_node
)
}
\description{
\code{BGe_score} Computes the BGe score of given network using precomputed sets of possible parents.
}
