% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{DAGcorescore}
\alias{DAGcorescore}
\title{BGe score}
\usage{
DAGcorescore(j, parentnodes, n, param)
}
\description{
\code{DAGcorescore} The log of the BGe score simplified as much as possible. This function is from BiDAG package.
}
