% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{edge_proposal}
\alias{edge_proposal}
\title{Markov Chain conventional single edge proposal move}
\usage{
edge_proposal(net, candidates, layers_def, ge_nodes, omics, B_prior_mat)
}
\description{
\code{edge_proposal} This function samples a conventional single edge proposal moves (identify those edges that are possible to change in given network structure)
}
