% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{edge_types}
\alias{edge_types}
\title{Resulting network definition}
\usage{
edge_types(PK, gene_annot, edge_list, node_list, OMICS_module_res)
}
\description{
\code{edge_types} Defines the resulting network structure and determines the color scale for each modality. This is part of trace_plots.
}
