% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{empB_heatmap}
\alias{empB_heatmap}
\title{Heatmap of empB - B}
\usage{
empB_heatmap(mcmc_res, OMICS_mod_res, gene_annot, TFtargs)
}
\description{
\code{empB_heatmap} plot a heatmap with empB - B values (depicts the difference between prior knowledge and the empirical knowledge)
}
