% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{energy_function_node_specific}
\alias{energy_function_node_specific}
\title{Node energy function}
\usage{
energy_function_node_specific(all_parents_config, B_prior_mat, int_node)
}
\description{
\code{energy_function_node_specific}  For each node returns its energy over all parent set configurations, the empty parent set is included.
}
