% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{epsilon}
\alias{epsilon}
\title{Epsilon}
\usage{
epsilon(net, B_prior_mat)
}
\description{
\code{epsilon} This function returns the epsilon value for each variable/node of the network. The sum of the epsilons of all variables/nodes in the network gives us the energy of given network.
}
