% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{fan_in_reverse}
\alias{fan_in_reverse}
\title{Number of reverse edge candidates}
\usage{
fan_in_reverse(positions, net_layer_max, layers_def)
}
\description{
\code{fan_in_reverse} Determine the number of edges that can be reversed using the fan-in restriction in the largest layer.
}
