% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{first_adapt_phase}
\alias{first_adapt_phase}
\title{1st adaption phase}
\usage{
first_adapt_phase(
  seed1,
  omics,
  B_prior_mat,
  energy_all_configs_node,
  len,
  layers_def,
  prob_mbr,
  BGe_score_all_configs_node,
  parent_set_combinations,
  annot
)
}
\description{
\code{first_adapt_phase} 1st adaption phase of the adaptive MCMC: the variance of the proposal distribution is changed to achieve the MC acceptance rate of 0.44.
}
