% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{flattenlist}
\alias{flattenlist}
\title{Flatten a list of lists}
\usage{
flattenlist(x)
}
\description{
\code{flattenlist} https://stackoverflow.com/questions/16300344/how-to-flatten-a-list-of-lists
}
