% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{init.net.mcmc}
\alias{init.net.mcmc}
\title{Random initial network}
\usage{
init.net.mcmc(omics, seed, layers_def)
}
\description{
\code{init.net.mcmc} This function is used to sample random initial network. The edges are sampled only between GE nodes.
}
