% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{lm_meth}
\alias{lm_meth}
\title{Linear regression GE~METH}
\usage{
lm_meth(ge_mat, meth_mat, gene, meth_probes, r_squared_thres)
}
\description{
\code{lm_meth} The linear regression model for a dependent variable GE and explanatory variable METH. Returns METH with significant coefficient, R^2 > threshold and R~Gaussian residuals.
}
