% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{MBR}
\alias{MBR}
\title{Markov Blanket Resampling}
\usage{
MBR(
  source_net_adjacency,
  layers_def,
  omics,
  BGe_score_all_configs_node,
  parent_set_combinations
)
}
\description{
\code{MBR} This function performs the markov blanket resampling method according to Su and Borsuk, 2016.
}
