% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{MC3}
\alias{MC3}
\title{Markov Chain conventional single edge proposal move}
\usage{
MC3(
  source_net_adjacency,
  omics,
  layers_def,
  B_prior_mat,
  beta.source,
  partition_func_UB_beta_source,
  parent_set_combinations,
  BGe_score_all_configs_node,
  annot
)
}
\description{
\code{MC3} This function samples a conventional single edge proposal move.
}
