% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{mcmc.simulation_sampling.phase}
\alias{mcmc.simulation_sampling.phase}
\title{Sampling phase}
\usage{
mcmc.simulation_sampling.phase(
  first,
  last,
  sim_init,
  prob_mbr,
  B_prior_mat,
  omics,
  parent_set_combinations,
  BGe_score_all_configs_node,
  layers_def,
  len,
  thin,
  counting,
  energy_all_configs_node
)
}
\description{
\code{mcmc.simulation_sampling.phase} This function performs the final sampling of network structures with estimated hyperparameters. It if part of sampling_phase function.
}
