% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{neighborhood_size}
\alias{neighborhood_size}
\title{Neighborhood size}
\usage{
neighborhood_size(net, layers_def, B_prior_mat, omics)
}
\description{
\code{neighborhood_size} This function is determines number of network structures that can be reached from the current network structure.
}
