% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{normalise}
\alias{normalise}
\title{Arrow of directed edges tuning}
\usage{
normalise(x, from = range(x), to = c(0, 1))
}
\description{
\code{normalise} This function is from the ambient package. It is used to determine the position of directed edge arrows.
}
