% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{parent_sets_sum_scores_childrenX}
\alias{parent_sets_sum_scores_childrenX}
\title{MBR sum of children scores}
\usage{
parent_sets_sum_scores_childrenX(
  parent_set_combinations,
  selected_node,
  children_selected_node,
  child_order,
  dag_tmp_bn,
  new_parent_set,
  source_net_adjacency,
  BGe_score_all_configs_node
)
}
\description{
\code{parent_sets_sum_scores_childrenX} This function determines the sum of BGe scores of given node's children.
}
