% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{parent_sets_sum_scores_X}
\alias{parent_sets_sum_scores_X}
\title{MBR sum of scores}
\usage{
parent_sets_sum_scores_X(
  parent_set_combinations,
  selected_node,
  descendants,
  parent_set,
  BGe_score_all_configs_node
)
}
\description{
\code{parent_sets_sum_scores_X} This function determines the sum of BGe scores of given node's parents.
}
