% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{pf_UB_est}
\alias{pf_UB_est}
\title{Partition function upper bound}
\usage{
pf_UB_est(omics, B_prior_mat, layers_def, annot)
}
\description{
\code{pf_UB_est} Partition function upper bound estimation with beta = 0. For each node returns energy over all possible parent set configurations and BGe score.
}
