% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{sample.chain}
\alias{sample.chain}
\title{Random initial network edge generation}
\usage{
sample.chain(empty_net, omics_ge, seed)
}
\description{
\code{sample.chain} This function is used to sample random initial network. The edges are sampled only between GE nodes.
}
