% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{sampling_phase}
\alias{sampling_phase}
\title{Sampling phase}
\usage{
sampling_phase(
  second.adapt.phase_net,
  seed1,
  seed2,
  omics,
  layers_def,
  prob_mbr,
  thin,
  minseglen,
  burn_in
)
}
\description{
\code{sampling_phase} Now we apply 2 MCMC simulations and check the RMS value. After the burn-in period, we discard the values from the first half of this phase.
}
