% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{scoreparameters_BiDAG_BGe}
\alias{scoreparameters_BiDAG_BGe}
\title{BGe score parameters}
\usage{
scoreparameters_BiDAG_BGe(n, data, bgepar = list(am = 1, aw = NULL))
}
\description{
\code{scoreparameters_BiDAG_BGe} Returns parameters needed for calculation of the BGe score. This function is from BiDAG package.
}
