% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{second_adapt_phase}
\alias{second_adapt_phase}
\title{Second adaption phase}
\usage{
second_adapt_phase(
  transient.phase_net,
  omics,
  layers_def,
  B_prior_mat,
  energy_all_configs_node,
  prob_mbr,
  BGe_score_all_configs_node,
  parent_set_combinations,
  annot
)
}
\description{
\code{second_adapt_phase} This phase identifies the proposal distribution that has a similar covariance structure with the target distribution.
}
