% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{source_net_def}
\alias{source_net_def}
\title{Source network for MCMC simulation}
\usage{
source_net_def(
  init.net.mcmc.output,
  omics,
  parent_set_combinations,
  BGe_score_all_configs_node,
  B_prior_mat,
  layers_def,
  energy_all_configs_node,
  len
)
}
\description{
\code{source_net_def} This function is used to create the initial network with its features necessary for MCMC simulation.
}
