% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{squared_jumping}
\alias{squared_jumping}
\title{Squared jumping of adaptive MCMC algorithm}
\usage{
squared_jumping(
  second.adapt.phase_net,
  constant,
  beta_sd,
  fin,
  B_prior_mat,
  omics,
  parent_set_combinations,
  BGe_score_all_configs_node,
  layers_def,
  prob_mbr,
  annot
)
}
\description{
\code{squared_jumping} Squared jumping of adaptive MCMC algorithm is used to tune the variance of the beta parameter.
}
