% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{trace_plots}
\alias{trace_plots}
\title{Trace plots of MCMC simulation & resulting network definition}
\usage{
trace_plots(
  mcmc_res,
  burn_in,
  thin,
  figures_dir,
  gene_annot,
  PK,
  OMICS_module_res,
  edge_freq_thres = NULL,
  gene_ID
)
}
\description{
\code{trace_plots} Create trace plots of MCMC simulation and filter low reliable edges based on the edge_freq_thres parameter. Defines the resulting network structure and determines the color scale for each modality.
}
