% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{transient_phase}
\alias{transient_phase}
\title{transient phase}
\usage{
transient_phase(
  first.adapt.phase_net,
  omics,
  B_prior_mat,
  layers_def,
  energy_all_configs_node,
  prob_mbr,
  BGe_score_all_configs_node,
  parent_set_combinations,
  annot
)
}
\description{
\code{transient_phase} This phase verify if the chain is moving towards the mode of target distribution.
}
