% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{variance_target}
\alias{variance_target}
\title{Second adaption phase}
\usage{
variance_target(
  transient.phase_net,
  constant,
  fin,
  B_prior_mat,
  omics,
  parent_set_combinations,
  BGe_score_all_configs_node,
  layers_def,
  prob_mbr,
  annot
)
}
\description{
\code{variance_target} This phase identifies the proposal distribution that has a similar covariance structure with the target distribution. This is part of second_adapt_phase.
}
