% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{B_prior_mat}
\alias{B_prior_mat}
\title{Biological prior matrix definition}
\usage{
B_prior_mat(omics, PK, layers_def, annot, r_squared_thres, lm_METH)
}
\description{
Biological prior matrix definition
}
