% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{OMICS_module}
\alias{OMICS_module}
\title{OMICS_module}
\usage{
OMICS_module(
  omics,
  PK,
  layers_def,
  annot = NULL,
  r_squared_thres = 0.3,
  lm_METH = TRUE
)
}
\description{
\code{OMICS_module} data preprocessing + B_prior_mat definition + partition function upper bound estimation + all possible parent sets per node definition + BGe score computation for all possible parent sets
}
